requirements.txt
curl_cffi
selectolax
python-dotenv